refactor: move deprecation warning tests to unit
